Replication, getting the server id would throw an exception sometimes. Backported to: 15.2.26.0;15.3.20.0 Reviewed by: Cameron
git-svn-id: https://opendentalsoft.com:23793/svn/opendental/opendental15.2@13045 098d0599-7f2b-0410-95a5-ad30ac85393a

+++ OpenDentBusiness/Data Interface/ReplicationServers.cs
@@ -84,7 +84,7 @@ namespace OpenDentBusiness{
 		///<summary>Gets the MySQL server_id variable for the current connection.</summary>
 		public static long GetServer_id() {
 			if(RemotingClient.RemotingRole==RemotingRole.ClientWeb) {
-				return Meth.GetInt(MethodBase.GetCurrentMethod());
+				return Meth.GetLong(MethodBase.GetCurrentMethod());
 			}
 			if(DataConnection.DBtype!=DatabaseType.MySql) {
 				return 0;
